+++ CHANGELOG.md
@@ -8,3 +8,4 @@
 - [go-kit](https://github.com/go-kit/kit/tree/master/log) logger constructor
 - [emperror](https://github.com/goph/emperror) error handler constructor
 - [airbrake](https://github.com/airbrake/gobrake) compatible error handler constructor
+- Error handler stack for collecting multiple handlers under a single dependency